requirements.txt
fastapi
uvicorn
gunicorn
pydantic
langchain
langchain-openai
python-dotenv
faiss-cpu
pypdf
langchain-community
selenium==4.15.2
webdriver-manager==4.0.1
openai